EasyConnectAPP/requirements.txt
PySide6>=6.8,<7.0
httpx>=0.27,<1.0
pyserial>=3.5,<4.0
esptool>=4.7,<5.0
Pillow>=10.3,<12.0
